refactor search, search result, app. WIP #12
